app/requirements.txt
streamlit
pyyaml
neo4j
chromadb
google-generativeai
streamlit-agraph # For interactive graph visualization
pandas
PySocks # Added for SOCKS proxy support
